feat: prevent default identity provider to be created
