+++ gravitee-am-management-api/gravitee-am-management-api-service/src/test/java/io/gravitee/am/management/service/impl/IdentityProviderManagerTest.java
@@ -22,13 +22,11 @@ import io.gravitee.am.model.ReferenceType;
 import io.gravitee.am.model.Role;
 import io.gravitee.am.plugins.idp.core.IdentityProviderPluginManager;
 import io.gravitee.am.service.RoleService;
-import io.gravitee.am.service.authentication.crypto.password.PasswordEncoderOptions;
 import io.reactivex.rxjava3.core.Flowable;
 import io.reactivex.rxjava3.core.Single;
 import io.reactivex.rxjava3.observers.TestObserver;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.jupiter.api.Assertions;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -36,7 +34,6 @@ import org.mockito.Spy;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.mock.env.MockEnvironment;
 
-import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
@@ -94,18 +91,14 @@ public class IdentityProviderManagerTest {
 
         verify(listener, times(2)).registerAuthenticationProvider(any());
 
-        verify(listener).registerAuthenticationProvider(argThat(idp -> {
-            return ReferenceType.ORGANIZATION.equals(idp.getReferenceType()) &&
-                    Organization.DEFAULT.equals(idp.getReferenceId()) &&
-                    idp.getRoleMapper() != null &&
-                    idp.getRoleMapper().containsKey("roleid");
-        }));
+        verify(listener).registerAuthenticationProvider(argThat(idp -> ReferenceType.ORGANIZATION.equals(idp.getReferenceType()) &&
+                Organization.DEFAULT.equals(idp.getReferenceId()) &&
+                idp.getRoleMapper() != null &&
+                idp.getRoleMapper().containsKey("roleid")));
 
-        verify(listener).registerAuthenticationProvider(argThat(idp -> {
-            return ReferenceType.ORGANIZATION.equals(idp.getReferenceType()) &&
-                    Organization.DEFAULT.equals(idp.getReferenceId()) &&
-                    idp.getType().equals("gravitee-am-idp");
-        }));
+        verify(listener).registerAuthenticationProvider(argThat(idp -> ReferenceType.ORGANIZATION.equals(idp.getReferenceType()) &&
+                Organization.DEFAULT.equals(idp.getReferenceId()) &&
+                idp.getType().equals("gravitee-am-idp")));
         verify(idpPluginManager).create(eq("gravitee-am-idp"), any(), any());
 
     }
@@ -153,55 +146,5 @@ public class IdentityProviderManagerTest {
         observer.assertValueCount(1);
 
         verify(idpPluginManager, never()).findById(any());
-    }
-
-    @Test
-    public void byDefault_DefaultIDPsPasswordAlgIsBCryptWith10IterationRounds() {
-        Map<String, Object> test = cut.createProviderConfiguration("test", null);
-
-        Assertions.assertEquals("BCrypt", test.get("passwordEncoder"));
-        Assertions.assertEquals(10, ((PasswordEncoderOptions) test.get("passwordEncoderOptions")).getRounds());
-    }
-
-    @Test
-    public void defaultIterationRoundsForBCryptIs10() {
-        environment.setProperty("domains.identities.default.passwordEncoder.algorithm", "BCrypt");
-
-        Map<String, Object> test = cut.createProviderConfiguration("test", null);
-
-        Assertions.assertEquals("BCrypt", test.get("passwordEncoder"));
-        Assertions.assertEquals(10, ((PasswordEncoderOptions) test.get("passwordEncoderOptions")).getRounds());
-    }
-
-    @Test
-    public void shouldReturnUpdatedIterationRoundsForBCryptAlg() {
-        environment.setProperty("domains.identities.default.passwordEncoder.algorithm", "BCrypt");
-        environment.setProperty("domains.identities.default.passwordEncoder.properties.rounds", "11");
-
-        Map<String, Object> test = cut.createProviderConfiguration("test", null);
-
-        Assertions.assertEquals("BCrypt", test.get("passwordEncoder"));
-        Assertions.assertEquals(11, ((PasswordEncoderOptions) test.get("passwordEncoderOptions")).getRounds());
-    }
-
-    @Test
-    public void defaultIterationRoundsForShaIs1() {
-        environment.setProperty("domains.identities.default.passwordEncoder.algorithm", "SHA-256");
-
-        Map<String, Object> test = cut.createProviderConfiguration("test", null);
-
-        Assertions.assertEquals("SHA-256", test.get("passwordEncoder"));
-        Assertions.assertEquals(1, ((PasswordEncoderOptions) test.get("passwordEncoderOptions")).getRounds());
-    }
-
-    @Test
-    public void shouldReturnUpdatedIterationRoundsForShaAlg() {
-        environment.setProperty("domains.identities.default.passwordEncoder.algorithm", "SHA-384");
-        environment.setProperty("domains.identities.default.passwordEncoder.properties.rounds", "11");
-
-        Map<String, Object> test = cut.createProviderConfiguration("test", null);
-
-        Assertions.assertEquals("SHA-384", test.get("passwordEncoder"));
-        Assertions.assertEquals(11, ((PasswordEncoderOptions) test.get("passwordEncoderOptions")).getRounds());
     }
 }
