+++ gravitee-am-management-api/gravitee-am-management-api-standalone/gravitee-am-management-api-standalone-distribution/src/main/resources/config/gravitee.yml
@@ -237,6 +237,7 @@ notifiers:
 domains:
 #  identities:
 #    default:
+#      enabled: false
 #      passwordEncoder:
 #         # Algorithms used to hash the user password.
 #         # Can be one of :
